refactor(ffmpeg): split transcode.go into stream.go, stream_video.go, stream_audio.go
- Extract stream interface and copyStreamState (with common inStream/outStream/frames
  fields) to stream.go; copyStreamState is now embedded by video and audio states
- Extract streamDecoder and videoStreamState to stream_video.go, inlining former
  videoEncoderState fields directly into videoStreamState
- Extract audioStreamState to stream_audio.go, inlining former audioEncoderState
  fields directly into audioStreamState
- Make receiveAndWritePackets a method on *copyStreamState, eliminating the
  framesPtr parameter
- Remove all else blocks: use early returns in configureEncoderPixelFormat (new
  helper extracted from openVideoEncoderContext) and continue in
  setupOutputContext's for loop
- Restructure channel-layout and sample-format selection in audioStreamState to
  use a default-value-then-override pattern instead of if/else
